LayoutTests/svg/dynamic-updates/SVGFESpotLightElement-dom-pointsAtX-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'pointsAtX' attribute of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS spotLightElement.getAttribute('pointsAtX') is "0"
PASS spotLightElement.getAttribute('pointsAtX') is "113"
PASS successfullyParsed is true

TEST COMPLETE
